--- tools/gfx/debug-layer.h
@@ -1,48 +1,0 @@
-
-
-namespace gfx
-{
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-class DebugCommandBuffer;
-
-
-
-
-
-class DebugTransientResourceHeap;
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-}
