+++ tools/gfx/debug-layer/debug-resource-views.cpp
@@ -1,0 +1,47 @@
+// debug-resource-views.cpp
+#include "debug-resource-views.h"
+
+namespace gfx
+{
+using namespace Slang;
+
+namespace debug
+{
+
+IResourceView::Desc* DebugResourceView::getViewDesc()
+{
+    SLANG_GFX_API_FUNC;
+
+    return baseObject->getViewDesc();
+}
+
+Result DebugResourceView::getNativeHandle(InteropHandle* outNativeHandle)
+{
+    SLANG_GFX_API_FUNC;
+
+    return baseObject->getNativeHandle(outNativeHandle);
+}
+
+DeviceAddress DebugAccelerationStructure::getDeviceAddress()
+{
+    SLANG_GFX_API_FUNC;
+
+    return baseObject->getDeviceAddress();
+}
+
+Result DebugAccelerationStructure::getNativeHandle(InteropHandle* outNativeHandle)
+{
+    SLANG_GFX_API_FUNC;
+
+    return baseObject->getNativeHandle(outNativeHandle);
+}
+
+IResourceView::Desc* DebugAccelerationStructure::getViewDesc()
+{
+    SLANG_GFX_API_FUNC;
+
+    return baseObject->getViewDesc();
+}
+
+} // namespace debug
+} // namespace gfx
