+++ tools/gfx/debug-layer/debug-sampler-state.cpp
@@ -1,0 +1,19 @@
+// debug-sampler-state.cpp
+#include "debug-sampler-state.h"
+
+namespace gfx
+{
+using namespace Slang;
+
+namespace debug
+{
+
+Result DebugSamplerState::getNativeHandle(InteropHandle* outNativeHandle)
+{
+    SLANG_GFX_API_FUNC;
+
+    return baseObject->getNativeHandle(outNativeHandle);
+}
+
+} // namespace debug
+} // namespace gfx
